internal/provider: add optional metadata param
This PR adds a new, optional metadata parameter, which controls whether
ScyllaDB Cloud client should fetch deployment metadata prior to issuing
API calls.

The deployment metadata is used to map various region, instance etc.
identifier to human-readable text strings.

By default metadata is always on and true.

It is safe to turn off the metadata retrieval, when it's known to
be not needed - e.g. for Stacks API integration.

For regular use of terraform providers user will not need to
control this behavior. The main use-case for this new param are
automation flows.